build: generate new HTML documentation
This revamped doc target converts and installs the remaining bits of
documentation mentioned in the sysusers manpages.

Markdown documentation is converted using cmark-gfm
if available, falling back to cmark or the original markdown.
When cmark-gfm is not found, rendering of GitHub Flavoured Markdown
features like tables is broken, but everything else is rendered fine.

The generated HTML pages don't look like the ones seen on systemd.io,
but I don't think it matters much; the important information is in the
text, not the graphics.

Addresses my previous comment,
https://github.com/cromerc/opensysusers/pull/1#discussion_r949653855